Make WeatherApi::new_url() only return a Url instead of Self
This is a fairly large change. It adds a few tests to check that the
urls are created correctly under different circumstances. The largest
change is that the API doesn't try to pretend it's a "class" anymore.
Instead, it uses the config to build a url and send that back.

** The following is a brain dump of my thoughts **

An idea I had is to perhaps make `fn new_url(config: &Config) -> Url`
instead return `Self`, with Self having a `url: Url` property in it.
That way the url can be mutable, and built upon with the builder
pattern.

I'm noticing the methods grow a lot as I extend the `Config`, so another
idea I had is to pass the config sections piecemeal at a time, so that
the `WeatherApi` trait can prescribe a builder pattern more easily. This
will make it easier to parse command line arguments with.